Refactor test and add explicit test for 1. dirty read 2. cascade delete
also update next version.md
